refactor: Updated presentation components and added presentation skeleton
