+++ src/app/lib/api.ts
@@ -1,10 +1,8 @@
 import axios, { AxiosRequestConfig } from "axios";
 import { getSession } from "next-auth/react";
 
-const API_URL = process.env.NEXT_PUBLIC_API_URL ?? 'http://localhost:5000/api';
-
 const axiosInstance = axios.create({
-    baseURL: API_URL,
+    baseURL: process.env.NEXT_PUBLIC_API_URL ?? 'http://localhost:5000/api',
     headers: {
         'Content-Type': 'application/json'
     }
